Fix compilation error in MoleculeTest
This fixes a compilation error in the MoleculeTest class
in which qCompare<unsigned int, unsigned long> was not
found.

+++ tests/auto/chemkit/molecule/moleculetest.cpp
@@ -381,27 +381,27 @@ void MoleculeTest::find()
 void MoleculeTest::rings()
 {
     chemkit::Molecule empty;
-    QCOMPARE(empty.rings().size(), 0UL);
+    QCOMPARE(empty.rings().size(), size_t(0));
     QCOMPARE(empty.ringCount(), 0);
 
     chemkit::Molecule cyclopropane;
     chemkit::Atom *cyclopropane_C1 = cyclopropane.addAtom("C");
     chemkit::Atom *cyclopropane_C2 = cyclopropane.addAtom("C");
     chemkit::Atom *cyclopropane_C3 = cyclopropane.addAtom("C");
-    QCOMPARE(cyclopropane.rings().size(), 0UL);
+    QCOMPARE(cyclopropane.rings().size(), size_t(0));
     QCOMPARE(cyclopropane.ringCount(), 0);
 
     cyclopropane.addBond(cyclopropane_C1, cyclopropane_C2);
     cyclopropane.addBond(cyclopropane_C2, cyclopropane_C3);
-    QCOMPARE(cyclopropane.rings().size(), 0UL);
+    QCOMPARE(cyclopropane.rings().size(), size_t(0));
     QCOMPARE(cyclopropane.ringCount(), 0);
 
     cyclopropane.addBond(cyclopropane_C1, cyclopropane_C3);
-    QCOMPARE(cyclopropane.rings().size(), 1UL);
+    QCOMPARE(cyclopropane.rings().size(), size_t(1));
     QCOMPARE(cyclopropane.ringCount(), 1);
 
     cyclopropane.removeAtom(cyclopropane_C2);
-    QCOMPARE(cyclopropane.rings().size(), 0UL);
+    QCOMPARE(cyclopropane.rings().size(), size_t(0));
     QCOMPARE(cyclopropane.ringCount(), 0);
 }
 
@@ -557,7 +557,7 @@ void MoleculeTest::conformers()
     chemkit::Atom *C1 = molecule.addAtom("C");
     chemkit::Atom *C2 = molecule.addAtom("C");
     QCOMPARE(molecule.conformerCount(), 1);
-    QCOMPARE(molecule.conformers().size(), 1UL);
+    QCOMPARE(molecule.conformers().size(), size_t(1));
     QVERIFY(molecule.conformer() != 0);
 
     chemkit::Conformer *conformer = molecule.addConformer();
